fix freelock queue
