+++ include/pump/poll/poller.h
@@ -24,7 +24,7 @@
 #include "pump/memory.h"
 #include "pump/net/socket.h"
 #include "pump/poll/channel.h"
-#include "pump/toolkit/freelock.h"
+#include "pump/toolkit/mutil_freelock_queue.h"
 
 namespace pump {
 namespace poll {
@@ -151,11 +151,11 @@ namespace poll {
 
         // Channel event
         std::atomic_int32_t cev_cnt_;
-        toolkit::freelock_list_queue<channel_event_ptr> cevents_;
+        toolkit::mutil_freelock_queue<channel_event_ptr> cevents_;
 
         // Channel tracker event
         std::atomic_int32_t tev_cnt_;
-        toolkit::freelock_list_queue<channel_tracker_event_ptr> tevents_;
+        toolkit::mutil_freelock_queue<channel_tracker_event_ptr> tevents_;
 
         // Channel trackers
         std::map<channel_tracker_ptr, channel_tracker_sptr> trackers_;
