+++ include/pump/service.h
@@ -19,8 +19,9 @@
 
 #include "pump/poll/poller.h"
 #include "pump/time/timer_queue.h"
-#include "pump/toolkit/features.h"
-#include "pump/toolkit/freelock.h"
+#include "pump/toolkit/mutil_freelock_queue.h"
+#include "pump/toolkit/single_freelock_queue.h"
+#include "pump/toolkit/block_freelock_queue.h"
 
 namespace pump {
 
@@ -126,12 +127,12 @@ namespace pump {
 
         // Posted task worker
         std::shared_ptr<std::thread> posted_task_worker_;
-        typedef toolkit::freelock_list_queue<post_task_type> post_task_impl_queue;
-        toolkit::block_freelock_queue<post_task_impl_queue> posted_tasks_;
+        typedef toolkit::mutil_freelock_queue<post_task_type> task_impl_queue;
+        toolkit::block_freelock_queue<task_impl_queue> posted_tasks_;
 
         // Timout timer worker
         std::shared_ptr<std::thread> timeout_timer_worker_;
-        typedef toolkit::single_freelock_list_queue<time::timer_wptr> timer_impl_queue;
+        typedef toolkit::single_freelock_queue<time::timer_wptr> timer_impl_queue;
         toolkit::block_freelock_queue<timer_impl_queue> timeout_timers_;
     };
     DEFINE_ALL_POINTER_TYPE(service);
