+++ include/pump/transport/tls_transport.h
@@ -17,9 +17,9 @@
 #ifndef pump_transport_tls_transport_h
 #define pump_transport_tls_transport_h
 
-#include "pump/toolkit/freelock.h"
 #include "pump/transport/flow/flow_tls.h"
 #include "pump/transport/base_transport.h"
+#include "pump/toolkit/mutil_freelock_queue.h"
 
 namespace pump {
 namespace transport {
@@ -172,7 +172,7 @@ namespace transport {
 
         // When sending data, transport will append buffer to sendlist at first. On
         // triggering send event, transport will send buffer in the sendlist.
-        toolkit::freelock_list_queue<toolkit::io_buffer_ptr> sendlist_;
+        toolkit::mutil_freelock_queue<toolkit::io_buffer_ptr> sendlist_;
     };
 
 }  // namespace transport
